+++ aiohttp_wsgi/__main__.py
@@ -4,7 +4,7 @@ Command line interface (CLI)
 
 If you don't need to add :ref:`websockets <aiohttp-web-websockets>` or
 :ref:`async request handlers <aiohttp-web-handler>` to your app, but still want to run your WSGI app on the
-:mod:`asyncio` event loop, :mod:`aiohttp_wsgi` provides a simpler command line interface.
+:mod:`asyncio` event loop, :mod:`aiohttp_wsgi` provides a simple command line interface.
 
 
 Example usage
@@ -37,10 +37,12 @@ You can view this reference at any time with ``aiohttp-wsgi-serve --help``.
 """
 
 import argparse
+import asyncio
 import logging
 import textwrap
 import aiohttp_wsgi
-from aiohttp_wsgi.api import serve, DEFAULTS, HELP
+from aiohttp_wsgi.api import start_server, DEFAULTS, HELP
+from aiohttp_wsgi.utils import parse_sockname
 
 
 parser = argparse.ArgumentParser(
@@ -61,7 +63,7 @@ def add_argument(name, *aliases, **kwargs):
         kwargs["help"] += " Can be specified multiple times."
     if kwargs["action"] == "count":
         kwargs.setdefault("default", 0)
-    if kwargs["action"] == "store":
+    if kwargs["action"] in ("append", "store"):
         kwargs.setdefault("default", DEFAULTS.get(varname))
         kwargs.setdefault("type", type(kwargs["default"]))
         assert not isinstance(None, kwargs["type"])
@@ -118,6 +120,9 @@ add_argument(
     "--max-request-body-size",
 )
 add_argument(
+    "--outbuf-overflow",
+)
+add_argument(
     "--verbose",
     "-v",
     action="count",
@@ -133,18 +138,37 @@ add_argument(
     "--version",
     action="version",
     help="Display version information.",
-    version="aiohttp-wsgi v{}".format(".".join(map(str, aiohttp_wsgi.__version__))),
+    version="aiohttp-wsgi v{}".format(aiohttp_wsgi.__version__),
 )
 
 
 def main():
     args = vars(parser.parse_args())
+    shutdown_timeout = args.pop("shutdown_timeout")
     # Set up logging.
     verbosity = (args.pop("verbose") - args.pop("quiet")) * 10
     logging.basicConfig(level=max(logging.ERROR - verbosity, logging.DEBUG), format="%(message)s")
     logging.getLogger("aiohttp").setLevel(max(logging.INFO - verbosity, logging.DEBUG))
     # Serve the app.
-    serve(**args)
+    loop = asyncio.get_event_loop()
+    server = loop.run_until_complete(start_server(loop=loop, **args))
+    server_uri = " ".join(
+        "http://{}:{}".format(*parse_sockname(socket.getsockname()))
+        for socket
+        in server.sockets
+    )
+    server.app.logger.info("Serving on %s", server_uri)
+    try:
+        loop.run_forever()
+    except KeyboardInterrupt:
+        pass
+    finally:
+        server.app.logger.debug("Waiting for server to shut down")
+        server.close()
+        server.app.logger.debug("Waiting for client connections to terminate")
+        loop.run_until_complete(server.wait_closed(shutdown_timeout=shutdown_timeout))
+        loop.close()
+        server.app.logger.info("Stopped serving on %s", server_uri)
 
 
 __doc__ = __doc__.format(**HELP, help=textwrap.indent(parser.format_help(), "    "))