+++ aiohttp_wsgi/api.py
@@ -1,71 +1,46 @@
-import asyncio
 import os
-import aiohttp
+from importlib import import_module
 from aiohttp.log import access_logger
 from aiohttp.web import Application, StaticRoute
 from aiohttp_wsgi.wsgi import WSGIHandler, DEFAULTS, HELP
-from aiohttp_wsgi.utils import parse_sockname, import_func
+from aiohttp_wsgi.utils import parse_sockname
 
 
 class Server:
 
-    def __init__(self, shutdown_timeout, app, handler, server, loop):
-        self._shutdown_timeout = shutdown_timeout
+    def __init__(self, app, handler, server):
         self.app = app
         self._handler = handler
         self._server = server
-        self._loop = loop
-        self._close_future = None
-        self._connector = None
-        self._session = None
 
     @property
     def sockets(self):
         return self._server.sockets
 
-    async def _wait_closed(self):
+    def close(self):
+        # Clean up unix sockets.
+        for socket in self.sockets:
+            host, port = parse_sockname(socket.getsockname())
+            if host == "unix":
+                os.unlink(port)
+        # Close the server.
+        self._server.close()
+
+    async def wait_closed(self, *, shutdown_timeout=60.0):
         await self._server.wait_closed()
         await self.app.shutdown()
-        await self._handler.finish_connections(self._shutdown_timeout)
+        await self._handler.finish_connections(shutdown_timeout)
         await self.app.cleanup()
 
-    def close(self):
-        if self._close_future is None:
-            # Clean up client session.
-            if self._session is not None:
-                self._session.close()
-            if self._connector is not None:
-                self._connector.close()
-            # Clean up unix sockets.
-            for socket in self.sockets:
-                host, port = parse_sockname(socket.getsockname())
-                if host == "unix":
-                    os.unlink(port)
-            # Close the server.
-            self._server.close()
-            self._close_future = self._loop.create_task(self._wait_closed())
 
-    async def wait_closed(self):
-        return await asyncio.shield(self._close_future, loop=self._loop)
-
-    async def __aenter__(self):
-        return self
-
-    async def __aexit__(self, *exc_info):
-        self.close()
-        await self.wait_closed()
-
-    async def request(self, method, path, **kwargs):
-        host, port = parse_sockname(self.sockets[0].getsockname())
-        if self._connector is None:
-            if host == "unix":
-                self._connector = aiohttp.UnixConnector(path=port, loop=self._loop)
-            else:
-                self._connector = aiohttp.TCPConnector(loop=self._loop)
-        if self._session is None:
-            self._session = aiohttp.ClientSession(connector=self._connector, loop=self._loop)
-        uri = "http://{}:{}{}".format(host, port, path)
-        return await self._session.request(method, uri, **kwargs)
+def import_func(func):
+    if isinstance(func, str):
+        assert ":" in func, "{!r} should have format 'module:callable'".format(func)
+        module_name, func_name = func.split(":", 1)
+        module = import_module(module_name)
+        func = getattr(module, func_name)
+    assert callable(func), "{!r} is not callable".format(func)
+    return func
 
 
 def format_path(path):
@@ -85,47 +60,34 @@ async def start_server(
     # Unix server config.
     unix_socket=None,
     unix_socket_perms=0o600,
-    # Prexisting socket config.
-    socket=None,
     # Shared server config.
     backlog=1024,
     # aiohttp config.
-    routes=(),
     static=(),
-    on_finish=(),
     # Asyncio config.
-    loop=None,
+    loop,
     # Aiohttp config.
     script_name="",
-    access_log=access_logger,
-    shutdown_timeout=60.0,
     **kwargs
 ):
-    loop = loop or asyncio.get_event_loop()
     app = Application(
         loop=loop,
     )
-    # Add routes.
-    for method, path, handler in routes:
-        app.router.add_route(method, path, import_func(handler))
     # Add static routes.
     for static_item in static:
-        if isinstance(static_item, str):
-            assert "=" in static_item, "{!r} should have format 'path=directory'"
-            static_item = static_item.split("=", 1)
+        assert "=" in static_item, "{!r} should have format 'path=directory'"
+        static_item = static_item.split("=", 1)
         path, dirname = static_item
         static_resource = app.router.add_resource("{}/{{filename:.*}}".format(format_path(path)))
         static_resource.add_route("*", StaticRoute(None, path + "/", dirname).handle)
-    # Add on finish callbacks.
-    for on_finish_callback in on_finish:
-        app.on_shutdown.append(import_func(on_finish_callback))
     # Add the wsgi application. This has to be last.
     app.router.add_route(
         "*",
         "{}{{path_info:.*}}".format(format_path(script_name)),
         WSGIHandler(import_func(application), loop=loop, **kwargs),
     )
-    handler = app.make_handler(access_log=access_log)
+    # HACK: Access logging is broken in aiohtp for unix sockets.
+    handler = app.make_handler(access_log=access_logger if unix_socket is None else None)
     # Set up the server.
     shared_server_kwargs = {
         "backlog": backlog,
@@ -134,12 +96,6 @@ async def start_server(
         server = await loop.create_unix_server(
             handler,
             path=unix_socket,
-            **shared_server_kwargs
-        )
-    elif socket is not None:
-        server = await loop.create_server(
-            handler,
-            sock=socket,
             **shared_server_kwargs
         )
     else:
@@ -153,33 +109,12 @@ async def start_server(
     if unix_socket is not None:
         os.chmod(unix_socket, unix_socket_perms)
     # All done!
-    return Server(shutdown_timeout, app, handler, server, loop)
-
-
-def serve(application, *, loop=None, **kwargs):  # pragma: no cover
-    loop = loop or asyncio.get_event_loop()
-    server = loop.run_until_complete(start_server(application, loop=loop, **kwargs))
-    server_uri = " ".join(
-        "http://{}:{}".format(*parse_sockname(socket.getsockname()))
-        for socket
-        in server.sockets
-    )
-    server.app.logger.info("Serving on %s", server_uri)
-    try:
-        loop.run_forever()
-    except KeyboardInterrupt:
-        pass
-    finally:
-        server.app.logger.debug("Waiting for server to shut down")
-        server.close()
-        server.app.logger.debug("Waiting for client connections to terminate")
-        loop.run_until_complete(server.wait_closed())
-        loop.close()
-        server.app.logger.info("Stopped serving on %s", server_uri)
+    return Server(app, handler, server)
 
 
 DEFAULTS = DEFAULTS.copy()
 DEFAULTS.update(start_server.__kwdefaults__)
+DEFAULTS.update(Server.wait_closed.__kwdefaults__)
 
 HELP = HELP.copy()
 HELP.update({